feat: moved header auth section to separate component
